+++ packages/chains/chains/58008.ts
@@ -47,6 +47,7 @@ export default {
       }
     ]
   },
+  "redFlags": [],
   "rpc": [
     "https://sepolia-pgn-public-goods-network.rpc.thirdweb.com/${THIRDWEB_API_KEY}",
     "https://58008.rpc.thirdweb.com/${THIRDWEB_API_KEY}",
@@ -54,5 +55,5 @@ export default {
   ],
   "shortName": "sepPGN",
   "slug": "sepolia-pgn-public-goods-network",
-  "testnet": false
+  "testnet": true
 } as const satisfies Chain;