+++ packages/react-native/src/evm/wallets/connectors/embedded-wallet/embedded/helpers/api/fetchers.ts
@@ -7,9 +7,9 @@ import {
   ROUTE_USER_MANAGED_OTP,
   ROUTE_VALIDATE_USER_MANAGED_OTP,
   ROUTE_IS_VALID_USER_MANAGED_OTP,
+  EWS_VERSION_HEADER,
 } from "../constants";
 import { getAuthTokenClient } from "../storage/local";
-import * as Application from "expo-application";
 import {
   RecoveryShareManagement,
   UserWalletStatus,
@@ -18,18 +18,26 @@ import {
   IsValidUserManagedEmailOTPResponse,
   VerifiedTokenResponse,
 } from "../../../types";
+import { createErrorMessage } from "../errors";
+import {
+  appBundleId,
+  reactNativePackageVersion,
+} from "../../../../../../utils/version";
+import { BUNDLE_ID_HEADER } from "../../../../../../constants/headers";
 
 const EMBEDDED_WALLET_TOKEN_HEADER = "embedded-wallet-token";
 const PAPER_CLIENT_ID_HEADER = "x-thirdweb-client-id";
-const BUNDLE_ID_HEADER = "x-bundle-id";
-const APP_BUNDLE_ID = Application.applicationId || "";
+const HEADERS = {
+  "Content-Type": "application/json",
+  [BUNDLE_ID_HEADER]: appBundleId,
+  [EWS_VERSION_HEADER]: reactNativePackageVersion,
+};
 
 export const verifyClientId = async (clientId: string) => {
   const resp = await fetch(ROUTE_VERIFY_THIRDWEB_CLIENT_ID, {
     method: "POST",
     headers: {
-      "Content-Type": "application/json",
-      [BUNDLE_ID_HEADER]: APP_BUNDLE_ID,
+      ...HEADERS,
     },
     body: JSON.stringify({ clientId, parentDomain: "" }),
   });
@@ -57,15 +65,17 @@ export const authFetchEmbeddedWalletUser = async (
         Authorization: `Bearer ${EMBEDDED_WALLET_TOKEN_HEADER}:${
           authTokenClient || ""
         }`,
-        [BUNDLE_ID_HEADER]: APP_BUNDLE_ID,
+        [BUNDLE_ID_HEADER]: appBundleId,
         [PAPER_CLIENT_ID_HEADER]: clientId,
+        [EWS_VERSION_HEADER]: reactNativePackageVersion,
       }
     : {
         Authorization: `Bearer ${EMBEDDED_WALLET_TOKEN_HEADER}:${
           authTokenClient || ""
         }`,
-        [BUNDLE_ID_HEADER]: APP_BUNDLE_ID,
+        [BUNDLE_ID_HEADER]: appBundleId,
         [PAPER_CLIENT_ID_HEADER]: clientId,
+        [EWS_VERSION_HEADER]: reactNativePackageVersion,
       };
   return fetch(url, params);
 };
@@ -117,8 +127,7 @@ export async function generateAuthTokenFromCognitoEmailOtp(
   const resp = await fetch(ROUTE_VERIFY_COGNITO_OTP, {
     method: "POST",
     headers: {
-      "Content-Type": "application/json",
-      [BUNDLE_ID_HEADER]: APP_BUNDLE_ID,
+      ...HEADERS,
     },
     body: JSON.stringify({
       access_token: session.getAccessToken().getJwtToken(),
@@ -142,8 +151,7 @@ export async function sendUserManagedEmailOtp(email: string, clientId: string) {
   const resp = await fetch(ROUTE_USER_MANAGED_OTP, {
     method: "POST",
     headers: {
-      "Content-Type": "application/json",
-      [BUNDLE_ID_HEADER]: APP_BUNDLE_ID,
+      ...HEADERS,
     },
     body: JSON.stringify({
       email,
@@ -168,8 +176,7 @@ export async function validateUserManagedEmailOtp(options: {
   const resp = await fetch(ROUTE_VALIDATE_USER_MANAGED_OTP, {
     method: "POST",
     headers: {
-      "Content-Type": "application/json",
-      [BUNDLE_ID_HEADER]: APP_BUNDLE_ID,
+      ...HEADERS,
     },
     body: JSON.stringify({
       email: options.email,
@@ -194,8 +201,7 @@ export async function isValidUserManagedEmailOtp(options: {
   const resp = await fetch(ROUTE_IS_VALID_USER_MANAGED_OTP, {
     method: "POST",
     headers: {
-      "Content-Type": "application/json",
-      [BUNDLE_ID_HEADER]: APP_BUNDLE_ID,
+      ...HEADERS,
     },
     body: JSON.stringify({
       email: options.email,
@@ -281,7 +287,7 @@ export async function getUserShares(clientId: string, getShareUrl: URL) {
     };
   } catch (e) {
     throw new Error(
-      `Malformed response from the ews user wallet API: ${JSON.stringify(e)}`,
+      createErrorMessage("Malformed response from the ews user wallet API", e),
     );
   }
 }