+++ packages/react/src/wallet/ConnectWallet/Modal/ConnectModalInline.tsx
@@ -1,8 +1,11 @@
 import { Cross2Icon } from "@radix-ui/react-icons";
 import { CrossContainer } from "../../../components/Modal";
 import { IconButton } from "../../../components/buttons";
-import { Theme, iconSize, radius, shadow } from "../../../design-system";
-import { WalletUIStatesProvider } from "../../../evm/providers/wallet-ui-states-provider";
+import { iconSize, radius, shadow } from "../../../design-system";
+import {
+  SetModalConfigCtx,
+  WalletUIStatesProvider,
+} from "../../../evm/providers/wallet-ui-states-provider";
 import {
   wideModalMaxHeight,
   modalMaxWidthCompact,
@@ -17,65 +20,26 @@ import {
   CustomThemeProvider,
   useCustomTheme,
 } from "../../../design-system/CustomThemeProvider";
+import { ComponentProps, useContext, useEffect } from "react";
+import { ConnectWalletProps } from "../ConnectWallet";
+import { useTWLocale } from "../../../evm/providers/locale-provider";
 import { StyledDiv } from "../../../design-system/elements";
-import { SyncedWalletUIStates } from "./ConnectEmbed";
-import { WelcomeScreen } from "../screens/types";
 
-/**
- * @internal
- */
-export type ConnectModalInlineProps = {
-  className?: string;
-  theme?: "dark" | "light" | Theme;
-
-  /**
-   * Set a custom title for the modal
-   * @defaultValue "Connect"
-   */
-  modalTitle?: string;
-
-  /**
-   * Replace the thirdweb icon next to modalTitle and set your own iconUrl
-   *
-   * Set to empty string to hide the icon
-   */
-  modalTitleIconUrl?: string;
-
-  style?: React.CSSProperties;
-
-  /**
-   * Set the size of the modal - `compact` or `wide` on desktop
-   *
-   * Modal size is always `compact` on mobile
-   *
-   * @defaultValue "wide"
-   */
-  modalSize?: "compact" | "wide";
-
-  /**
-   * If provided, Modal will show a Terms of Service message at the bottom with below link
-   */
-  termsOfServiceUrl?: string;
-
-  /**
-   * If provided, Modal will show a Privacy Policy message at the bottom with below link
-   */
-  privacyPolicyUrl?: string;
-
-  /**
-   * Customize the welcome screen
-   *
-   * Either provide a component to replace the default screen entirely
-   *
-   * or an object with title, subtitle and imgSrc to change the content of the default screen
-   */
-  welcomeScreen?: WelcomeScreen;
-};
-
-/**
- * @internal
- */
-export const ConnectModalInline = (props: ConnectModalInlineProps) => {
+export const ConnectModalInline = (
+  props: Omit<
+    ConnectWalletProps,
+    | "detailsBtn"
+    | "dropdownPosition"
+    | "auth"
+    | "networkSelector"
+    | "hideTestnetFaucet"
+    | "switchToActiveChain"
+    | "supportedTokens"
+    | "hideSwitchToPersonalWallet"
+  > & {
+    onModalHide?: () => void;
+  },
+) => {
   const { screen, setScreen, initialScreen } = useScreen();
   const walletConfigs = useWallets();
   const modalSize =
@@ -89,7 +53,9 @@ export const ConnectModalInline = (props: ConnectModalInlineProps) => {
         screen={screen}
         setScreen={setScreen}
         setHideModal={() => {
-          // no op
+          if (props.onModalHide) {
+            props.onModalHide();
+          }
         }}
       />
 
@@ -129,10 +95,9 @@ export const ConnectModalInline = (props: ConnectModalInlineProps) => {
               modalSize === "compact"
                 ? modalMaxWidthCompact
                 : modalMaxWidthWide,
-            ...props.style,
           }}
         >
-          {modalSize === "compact" ? (
+          {props.modalSize === "compact" ? (
             <DynamicHeight> {content} </DynamicHeight>
           ) : (
             content
@@ -143,6 +108,39 @@ export const ConnectModalInline = (props: ConnectModalInlineProps) => {
     </WalletUIStatesProvider>
   );
 };
+
+function SyncedWalletUIStates(
+  props: ComponentProps<typeof WalletUIStatesProvider>,
+) {
+  const setModalConfig = useContext(SetModalConfigCtx);
+  const locale = useTWLocale();
+
+  // update modalConfig on props change
+  useEffect(() => {
+    setModalConfig((c) => ({
+      ...c,
+      title: props.title || locale.connectWallet.defaultModalTitle,
+      theme: props.theme || "dark",
+      modalSize: (isMobile() ? "compact" : props.modalSize) || "wide",
+      termsOfServiceUrl: props.termsOfServiceUrl,
+      privacyPolicyUrl: props.privacyPolicyUrl,
+      welcomeScreen: props.welcomeScreen,
+      titleIconUrl: props.titleIconUrl,
+    }));
+  }, [
+    props.title,
+    props.theme,
+    props.modalSize,
+    props.termsOfServiceUrl,
+    props.privacyPolicyUrl,
+    props.welcomeScreen,
+    props.titleIconUrl,
+    setModalConfig,
+    locale.connectWallet.defaultModalTitle,
+  ]);
+
+  return <WalletUIStatesProvider {...props} />;
+}
 
 const ConnectModalInlineContainer = /* @__PURE__ */ StyledDiv(() => {
   const theme = useCustomTheme();