+++ packages/wallets/src/evm/connectors/injected/types.ts
@@ -52,6 +52,8 @@ type InjectedProviderFlags = {
   isZerion?: true;
   isOkxWallet?: true;
   isCoreWallet?: true;
+  isDefiWallet?: true;
+  isRabbyWallet?: true;
   isCoin98Wallet?: true;
 };
 type InjectedProviders = InjectedProviderFlags & {