+++ packages/wallets/src/evm/index.ts
@@ -53,6 +53,15 @@ export { getInjectedOKXProvider } from "./connectors/okx/getInjectedOKXProvider"
 export { CoreWallet, type CoreWalletOptions } from "./wallets/core-wallet";
 export { getInjectedCoreWalletProvider } from "./connectors/core-wallet/getInjectedCoreWalletProvider";
 
+export {
+  CryptoDefiWallet,
+  type CryptoDefiWalletOptions,
+} from "./wallets/crypto-defi-wallet";
+export { getInjectedCryptoDefiWalletProvider } from "./connectors/crypto-defi-wallet/getInjectedCryptoDefiWalletProvider";
+
+export { RabbyWallet, type RabbyOptions } from "./wallets/rabby";
+export { getInjectedRabbyProvider } from "./connectors/rabby/getInjectedRabbyProvider";
+
 export { Coin98Wallet, type Coin98Options } from "./wallets/coin98";
 export { getInjectedCoin98Provider } from "./connectors/coin98/getInjectedCoin98Provider";